Add a periodic job for the k8s-cluster-api-provider project
A periodic job for the k8s-cluster-api-provider project,
is intended for execution on specific branches.
Find the job definition by referring to this link: https://github.com/SovereignCloudStack/k8s-cluster-api-provider/blob/main/.zuul.yaml

By doing so, we have the capability to specify and manage
the branches on which the periodic job will be run through Zuul.

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>